Migrated server from using SyncMaps to manually locking maps
This is part a proposed revision to the assignment. Previously,
students were instructed to use the provided SyncMap as a way of
averting concurrent reads/writes to maps. However, this data type
is limiting in how it allows students to manage concurrent reads/
writes, so instead this proposes a solution that manually locks
maps (and other fields) when concurrent reads/writes are possible.