Parse structs in dbg.LogKVs so we can see field names in log
By default, Go converts a struct to a string that only contains
the values of the fields. This is not very useful for logging,
since we cannot see the field names and figure out which field is
which based on their ordering. I added to dbg so that it parses
structs manually and prints the field names and values to the
log.