Added check for convert to follower signal at top of beLeader
Since beLeader grabs a lock at the beginning, it could be blocked
for a short time, during which a convert to follower signal
could be sent. Ideally, we would want to "select" between getting
the lock and getting this signal, and execute based on which
happens first. But since we cannot select on grabbing a lock,
we just check after getting it whether a signal was sent, and
convert to follower if so. Otherwise it's safe to proceed as
leader.