Implemented better fix for deadlock bug from a few commits ago
The last fix wasn't really a fix - it was more of a hack that
eliminated deadlock by just not locking (which isn't correct).
Instead, the solution was to actually increase the scope of the
locked portions of code so they can't overlap: from the decision
to send a CTF signal, through the end of conversion. In particular,
we prevent sending of an "outdated" signal by:
- When deciding to send a CTF signal, always compare terms against
  rf.CurrentTerm (the latest actual term)
- Lock around this comparison, all the way through to the end of
  conversion (i.e. end of sendConvertToFollowerSig()). This ensures
  that the signal+conversion happen as an atomic unit, so we
  can't send a signal based on an outdated term.
- Because of this, we also were able to remove currentTerm as a
  field in the signal, since this will always send a signal
  relevant to the current term.