Fixed bug causing deadlock upon receiving two CTF sigs as leader
The following scenario was possible: You are a leader and you
receive a CTF signal, causing you to restart beLeader from the
top. beLeader has to grab a lock before proceeding. However, if
you get an AE request or reply that causes another CTF signal to
be sent, and it gets the lock first, then you cannot proceed in
beLeader to read new CTF signals, and you end up in deadlock.

The core problem here is that beLeader must wait for the lock
before it can check for CTF signals. But really we need to
"select" between these two cases - whichever happens first should
determine what to execute. Go's locking mechanism doesn't let
you select on grabbing a lock though, so I implemented a
workaround that achieves the same effect.