Refactored state transitions
State transitions used to be handled by having each state
handler spin off a goroutine for the next state when triggered
to change states. Although this worked and was fairly intuitive,
it was not ideal because it did not enforce correctness in the
structure of the code - e.g. it was possible to have two states
running at the same time if one state handler did not properly
return after starting another one.

Since I believe it is a good practice, I decided to revise this
(even though it was already correct) so that the structure of the
code enforced correctness - that no two states can be running at
once. The functions for each state now return the next state to
transition to, and one parent function, run(), then makes the
transition to the next state.