Send entries to applyCh without holding a lock
I hold a lock while applying entries, and this included while
sending on the applyCh. However, there is a possiblity that the
applyCh could fill up and block, which would prevent us from
returning the lock. This hasn't actually happened, but I decided
to fix it anyway. Now I only hold the lock to compute the entries
that should be applied, and then send them to applyCh after
returning the lock.