Changed timer intervals
I would fail the "unreliable" tests moderately often, and it
always seemed to be because the servers could not reach agreement
for a few commands. I suspect that this is because they are not
reaching agreement quickly enough for the test's time constraints,
as was the case for TestBackup (see several commits ago). So I
halved the leader's AppendEntries interval, effectively making
consensus twice as fast. I also changed the Apply interval so
that it wasn't an even multiple of the leader interval, making
it less likely for the two to compete for a lock.

Lastly, I changed the RPC timeout to 50ms as recommended by
Andrew Or on Piazza last year. After a timing experiment, I
found that most AppendEntries RPCs (in unreliable mode) take
<25ms, with some occasionally taking 30-45ms. So setting the
timeout at 50ms catches almost all cases, and over 50ms we can
safely assume the server will not be responding. This introduced
the possibility that two RPC replies from different rounds of
AppendEntries could be processed simultaneously, so I fixed
the nextIndex/matchIndex update code to be concurrency-safe.