Allowed dbg.LogKVs to ignore non-exported struct fields
Before this commit, if a struct was passed as a value to
dbg.LogKVs, it was required to have all its fields be exported,
or nothing would print and an error would be returned. This was
an intentional design decision since I believe it is a bad
design practice to silently fail to print. However, this was
limiting, since I use LogKVs frequently to print the Raft object,
and it takes a lot of space to print some variables (like the
persister) that are unnecessary. So I compromised and allowed
dbg to ignore the unexported fields and print only the exported
ones, but it still returns an error to indicate that unexported
fields were ignored. Also, I modified raft.go to unexport fields
that I don't need to print, so I save lots of space in the log
output.

Also, I sort the keys passed to LogKVs so they print in a
consistent order. This always bugged me when reading log output.