Fixed bug causing slow elections in unreliable network
Turns out that there was a pretty major bug in my implementation
of the protocol, regarding elections. I interpreted "convert to
follower" to also include resetting the election timer, but these
are actually independent. The timer is actually global and carries
across the follower and candidate states, and it gets reset in
3 places: upon receiving AE from valid leader, upon granting a vote
in RV, and upon entering the candidate state. Sometimes it
corresponds to converting to follower (e.g. upon receiving AE),
but not always, so they should not be coupled. I fixed this by
using a separate signal to indicate that the timer should be reset.