Fixed bug causing seg faults in dbg statements
I haven't been able to confirm exactly what the issue is, but I
speculate that when calling dbg.LogKVs with an array as a kv,
for instance, if we remove entries from the array concurrently
to dbg processing the array, dbg might try to access an entry
that was deleted, causing a seg fault. Unforunately, there is no
way for dbg to prevent this internally, so the only way is for
the caller to ensure that such an array is not modified
concurrently to the dbg.LogKVs call.

I believe this would happen with rf.Log, so I got around it
by only printing the log if I am holding a lock.