Fixed bug causing a server to never run in an election
There was a subtle scenario in which a server's election timer
wouldn't get reset, causing him to get stuck as a follower
forever: If a follower's election timer times out, he will
transition to candidate. But in order to reach the statement where
he resets the election timer, he must select between grabbing a
lock and receiving a CTF signal. If he receives the CTF signal
first, then he'll transition back to follower, *without* having
reset the timer. Included logs of this scenario.

Fixed this by resetting the timer before that select statement.