Cleaned up potentially confusing logic in beFollower
beFollower would follow the convention of checking for CTF signals
before actually entering the state (like candidate and leader
states). However, this is unnecessary, as it's impossible to
receive a CTF signal at this point: the only way to transition to
follower (after initializing the server) is via a CTF signal.
But because we are holding a lock in every place we send a CTF
signal, and we hold it until the conversion is complete (via
waiting for an ack) it's impossible to send two at the same time.
Thus we cannot receive a second signal until after the ack has
been sent. So I removed this unnecessary logic at the beginning
of beFollower.