Fixed bug causing deadlock after restarting a server
If, immediately after restarting a server with a call to Make(),
a server receives an RPC (e.g. AppendEntries), the server might
end up in deadlock. Make spins off the main routine run(), which
then puts the server into the follower state, at which point it
is ready to accept CTF signals. But if the server receives an
AppendEntries first, and it grabs the lock, then it may send a
CTF signal (waiting for an ack before continuing), and beFollower
is not yet receiving CTF signals because it hasn't obtained the
lock yet. Included logs demonstrating this problem.

Fixed this by having Make() lock until the server is in the
follower state, and thus ready to accept CTF signals.